Convert the Rule Audit pages to use PageRoutedTabs. Add Back to Rule Audit tab.
